ALl getAll, getById and Post done

+++ src/main/java/org/carRental/dao/CustomerDAO.java
@@ -7,7 +7,6 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.ArrayList;
 import java.util.List;
 
 import static org.carRental.dao.CustomerSqlQueryConstant.*;
@@ -48,7 +47,7 @@ public class CustomerDAO extends BaseDAO implements ICrud<Customer> {
     public Customer getById(Integer id) {
         try {
             PreparedStatement ps = conn.prepareStatement(GET_BY_ID);
-            ps.setInt(1, id.intValue());
+            ps.setInt(1, id);
             ResultSet res = ps.executeQuery();
             return customerMapper.resultToObject(res);
         } catch (SQLException e) {
@@ -71,7 +70,7 @@ public class CustomerDAO extends BaseDAO implements ICrud<Customer> {
         try {
             PreparedStatement ps = conn.prepareStatement(UPDATE_NAME_BY_ID);
             ps.setString(1, obj.getCustomer_name());
-            ps.setInt(2, id.intValue());
+            ps.setInt(2, id);
             ps.executeUpdate();
         } catch (SQLException e) {
             throw new RuntimeException(e);
@@ -83,7 +82,7 @@ public class CustomerDAO extends BaseDAO implements ICrud<Customer> {
     public void delete(Integer id) {
         try {
             PreparedStatement ps = conn.prepareStatement(DELETE);
-            ps.setInt(1, id.intValue());
+            ps.setInt(1, id);
             ps.executeUpdate();
         } catch (SQLException e) {
             throw new RuntimeException(e);
